rename httpclient to client
