+++ pkg/platform/abstract/project/external/leader/client/client.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package httpclient
+package client
 
 import (
 	"context"
